+++ collector-amazon-ivs/src/main/java/com/bitmovin/analytics/amazon/ivs/player/IvsPositionProvider.kt
@@ -1,8 +1,0 @@
-package com.bitmovin.analytics.amazon.ivs.player
-
-import com.amazonaws.ivs.player.Player
-
-internal class IvsPositionProvider(private val player: Player) : PositionProvider {
-    override val position: Long
-        get() = player.position
-}