+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/player/BitmovinPlayerContext.kt
@@ -1,0 +1,14 @@
+package com.bitmovin.analytics.bitmovin.player.player
+
+import com.bitmovin.analytics.adapters.PlayerContext
+import com.bitmovin.analytics.bitmovin.player.BitmovinUtil
+import com.bitmovin.player.api.Player
+
+internal class BitmovinPlayerContext(private val player: Player) : PlayerContext {
+    override fun isPlaying(): Boolean {
+        return player.isPlaying
+    }
+
+    override val position: Long
+        get() = BitmovinUtil.getCurrentTimeInMs(player)
+}