+++ collector-bitmovin-player/src/test/java/com/bitmovin/analytics/bitmovin/player/BitmovinSdkAdapterTest.kt
@@ -42,8 +42,8 @@ class BitmovinSdkAdapterTest {
             PlayerStateMachine(
                 mockk(),
                 mockk(),
-                mockk(),
                 qualityChangeEventLimiter,
+                mockk(),
                 mockk(),
                 mockk(),
             ),