+++ collector-exoplayer/src/test/java/com/bitmovin/analytics/exoplayer/ExoPlayerAdapterTest.kt
@@ -34,8 +34,8 @@ class ExoPlayerAdapterTest {
             PlayerStateMachine(
                 mockk(),
                 mockk(),
-                mockk(),
                 qualityChangeEventLimiter,
+                mockk(),
                 mockk(),
                 mockk(),
             ),